requirements.txt
langchain
langchain-community==0.2.9
langchain-core
fastapi==0.111.1
langserve==0.2.2
langchain-text-splitters==0.2.2
python-dotenv==1.0.1
sse_starlette==2.1.2
langchain-google-genai
google-generativeai
langchain_chroma